spec: clarify language for data types
