+++ spec/000-introduction.md
@@ -1,3 +1,9 @@
-# Introduction
+# The Hyper Framework Specification
 
---write this--
+This is a specification for the hyper backend framework. For more information, see [github.com/hntrl/hyper](https://github.com/hntrl/hyper)
+
+Hyper is a DSL (domain specific language) built for developing systems utilizing reactive architecture, conforming with [domain-driven design](https://en.wikipedia.org/wiki/Domain-driven_design), and designed with distributed computing in mind. It is interpreted, strongly typed, and has explicit support for separating business logic from the details of its deployment.
+
+Hyper differs from a typical programming language in the sense that it does not follow a standard control flow. Instead, it is designed to be a tool for describing the capabilities of a system in a way that is easy to interpret, and then provides a functional paradigm to describe the control flow of those capabilities. This allows for a system to be described in a way that is easy to understand and reason about, while still providing the flexibility to be deployed in a variety of ways.
+
+<!-- This allows for tooling in every stage of the development life cycle to allow for consistent implementations and informed communication between teams across multiple disciplines.-->
